Renamed IOSAudio and IOSInput implementation to follow naming convention
